BJS2-77677 during PR review refactoring
