+++ src/main/java/faang/school/analytics/mapper/event/PostServiceEventMapper.java
@@ -1,0 +1,4 @@
+package faang.school.analytics.mapper.event;
+
+public interface PostServiceEventMapper {
+}
