--- src/main/java/faang/school/analytics/mapper/goalcompleted/PostServiceEventMapper.java
@@ -1,4 +1,0 @@
-package faang.school.analytics.mapper.goalcompleted;
-
-public interface PostServiceEventMapper {
-}
