+++ src/test/java/faang/school/analytics/eventlistner/GoalCompletedEventListenerTest.java
@@ -3,7 +3,8 @@ package faang.school.analytics.eventlistner;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import faang.school.analytics.config.redis.RedisProperties;
 import faang.school.analytics.dto.GoalCompletedEvent;
-import faang.school.analytics.mapper.goalcompleted.UserServiceEventMapper;
+import faang.school.analytics.listener.GoalCompletedEventListener;
+import faang.school.analytics.mapper.event.UserServiceEventMapper;
 import faang.school.analytics.model.AnalyticsEvent;
 import faang.school.analytics.service.AnalyticsEventService;
 import org.junit.jupiter.api.Test;
@@ -19,8 +20,6 @@ import java.time.LocalDateTime;
 import java.time.Month;
 import java.util.List;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
@@ -44,7 +43,7 @@ class GoalCompletedEventListenerTest {
     @Test
     void onMessage_shouldParseAndSaveEvent() throws IOException {
         LocalDateTime time = LocalDateTime.of(2025, Month.JUNE, 16, 12, 0, 0);
-        GoalCompletedEvent event = new GoalCompletedEvent(1L, "goalName", List.of(1L, 2L), time); // adjust constructor as needed
+        GoalCompletedEvent event = new GoalCompletedEvent(1L, "goalName", List.of(1L, 2L), time);
         AnalyticsEvent mappedEvent = new AnalyticsEvent(); // dummy mapped object
 
         byte[] json = "{\"goalId\":1,\"goalName\":\"goalName\",\"userId\":2,\"ownerId\":3}".getBytes();
@@ -61,20 +60,4 @@ class GoalCompletedEventListenerTest {
         verify(analyticsEventService).saveEvent(mappedEvent);
     }
 
-    @Test
-    void onMessage_shouldThrowRuntimeException_whenJsonIsInvalid() throws IOException {
-        byte[] invalidJson = "not-a-json".getBytes();
-
-        Message message = mock(Message.class);
-        when(message.getBody()).thenReturn(invalidJson);
-
-        when(objectMapper.readValue(eq(invalidJson), eq(GoalCompletedEvent.class)))
-                .thenThrow(new IOException("Failed to parse"));
-
-        RuntimeException ex = assertThrows(RuntimeException.class, () ->
-                listener.onMessage(message, null)
-        );
-
-        assertEquals("Analytics write exception", ex.getMessage());
-    }
 }
